+++ sheet01/identification-Q234/normalized_hist.m
@@ -14,7 +14,7 @@ function h = normalized_hist(img_gray, num_bins)
     
     assert(num_bins <= 255, 'num_bins is not allowed to be bigger than 255');
     assert(num_bins > 0, 'num_bins has to be > 0');
-    num_bins = uint8(num_bins);
+    num_bins = num_bins;
   
     % data is collecting the amout of times a specific range of values occurs
     % the range is determined by num_bins which specifies the amout of ranges
